+++ trail/app/src/main/java/com/example/trail/NewTask/SimpleTask/MyLocation.java
@@ -27,4 +27,8 @@ public class MyLocation implements Serializable {
     public MyLocation(String s){
         place = s;
     }
+    @Override
+    public String toString(){
+        return place+"|"+latitude+"|"+longitude;
+    }
 }
