+++ trail/app/src/main/java/com/example/trail/Setting/AssignWithML/AssignActivity.java
@@ -1,0 +1,390 @@
+package com.example.trail.Setting.AssignWithML;
+
+import android.graphics.RectF;
+import android.os.Bundle;
+import android.util.TypedValue;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.widget.LinearLayout;
+import android.widget.RelativeLayout;
+import android.widget.TextView;
+import android.widget.Toast;
+
+import androidx.appcompat.app.AlertDialog;
+
+import com.alamkanak.weekview.DateTimeInterpreter;
+import com.alamkanak.weekview.MonthLoader;
+import com.alamkanak.weekview.WeekView;
+import com.alamkanak.weekview.WeekViewEvent;
+import com.example.trail.MainActivity;
+import com.example.trail.NewTask.Collection.TaskCollector;
+import com.example.trail.NewTask.SimpleTask.Task;
+import com.example.trail.R;
+import com.example.trail.Utility.Utils.TimeUtil;
+
+import java.text.SimpleDateFormat;
+import java.util.ArrayList;
+import java.util.Calendar;
+import java.util.List;
+import java.util.Locale;
+
+import solid.ren.skinlibrary.base.SkinBaseActivity;
+
+public class AssignActivity extends SkinBaseActivity implements WeekView.EventClickListener, MonthLoader.MonthChangeListener, WeekView.EventLongPressListener, WeekView.EmptyViewLongPressListener {
+    private static final int TYPE_DAY_VIEW = 1;
+    private static final int TYPE_THREE_DAY_VIEW = 2;
+    private static final int TYPE_WEEK_VIEW = 3;
+    private int mWeekViewType = TYPE_THREE_DAY_VIEW;
+    private LinearLayout back;
+    private TextView title;
+    private WeekView mWeekView;
+    private AlertDialog dialog;
+    private List<WeekViewEvent> events = new ArrayList<WeekViewEvent>();
+    private List<AssignHelper> helpers = new ArrayList<>();
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_assign);
+        back = findViewById(R.id.back);
+        title = findViewById(R.id.title);
+        back.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                AssignActivity.this.finish();
+            }
+        });
+        title.setText(getResources().getString(R.string.assign));
+        try {
+            long id = 0;
+            for (TaskCollector taskCollector: MainActivity.taskCollectors){
+                for (Task task:taskCollector.getTasks()){
+                    if (task.getExpireTime()!=null&&task.getRemindTime()==null){
+                        AssignHelper helper = new AssignHelper(task,null);
+                        helper.createInlineSource(MainActivity.taskCollectors);
+                        helpers.add(helper);
+                    }
+                    else if(task.getExpireTime()!=null&&task.getRemindTime()!=null){
+                        WeekViewEvent event = new WeekViewEvent(++id, task.getTitle(), TimeUtil.Date2Cal(task.getRemindTime()), TimeUtil.Date2Cal(task.getExpireTime()));
+                        switch (task.getPriority()) {
+                            case EUGEN:
+                                event.setColor(R.color.event_color_01);
+                                break;
+                            case MIDDLE:
+                                event.setColor(R.color.event_color_02);
+                                break;
+                            case LITTLE:
+                                event.setColor(R.color.event_color_03);
+                                break;
+                            default:
+                                event.setColor(R.color.event_color_04);
+                                break;
+                        }
+                        events.add(event);
+                    }
+                }
+            }
+            for (AssignHelper helper:helpers){
+                if (helper.getPredictTime()!=null) {
+                    String time = helper.getPredictTime();
+                    Task task = helper.getMTask();
+                    if (time.length() != 10) {
+                        Calendar calendar = Calendar.getInstance();
+                        calendar.setTimeInMillis(Long.valueOf(time.substring(0, time.length() - 3)));
+                        WeekViewEvent event = new WeekViewEvent(++id, task.getTitle(), calendar, TimeUtil.Date2Cal(task.getExpireTime()));
+                        switch (task.getPriority()) {
+                            case EUGEN:
+                                event.setColor(R.color.event_color_01);
+                                break;
+                            case MIDDLE:
+                                event.setColor(R.color.event_color_02);
+                                break;
+                            case LITTLE:
+                                event.setColor(R.color.event_color_03);
+                                break;
+                            default:
+                                event.setColor(R.color.event_color_04);
+                                break;
+                        }
+                        events.add(event);
+                    }
+                }
+            }
+        } catch (Exception e){
+            e.printStackTrace();
+        }
+
+        // Get a reference for the week view in the layout.
+        mWeekView = (WeekView) findViewById(R.id.weekView);
+
+        // Show a toast message about the touched event.
+        mWeekView.setOnEventClickListener(this);
+
+        // The week view has infinite scrolling horizontally. We have to provide the events of a
+        // month every time the month changes on the week view.
+        mWeekView.setMonthChangeListener(this);
+
+        // Set long press listener for events.
+        mWeekView.setEventLongPressListener(this);
+
+        // Set long press listener for empty view
+        mWeekView.setEmptyViewLongPressListener(this);
+
+        // Set up a date time interpreter to interpret how the date and time will be formatted in
+        // the week view. This is optional.
+        setupDateTimeInterpreter(false);
+    }
+
+    /**
+     * Set up a date time interpreter which will show short date values when in week view and long
+     * date values otherwise.
+     * @param shortDate True if the date values should be short.
+     */
+    private void setupDateTimeInterpreter(final boolean shortDate) {
+        mWeekView.setDateTimeInterpreter(new DateTimeInterpreter() {
+            @Override
+            public String interpretDate(Calendar date) {
+                SimpleDateFormat weekdayNameFormat = new SimpleDateFormat("EEE", Locale.getDefault());
+                String weekday = weekdayNameFormat.format(date.getTime());
+                SimpleDateFormat format = new SimpleDateFormat(" M/d", Locale.getDefault());
+
+                // All android api level do not have a standard way of getting the first letter of
+                // the week day name. Hence we get the first char programmatically.
+                // Details: http://stackoverflow.com/questions/16959502/get-one-letter-abbreviation-of-week-day-of-a-date-in-java#answer-16959657
+                if (shortDate)
+                    weekday = String.valueOf(weekday.charAt(0));
+                return weekday.toUpperCase() + format.format(date.getTime());
+            }
+
+            @Override
+            public String interpretTime(int hour) {
+                return hour > 11 ? (hour - 12) + " PM" : (hour == 0 ? "12 AM" : hour + " AM");
+            }
+        });
+    }
+
+    @Override
+    public void onEventClick(WeekViewEvent event, RectF eventRect) {
+        Toast.makeText(this, "Clicked " + event.getName(), Toast.LENGTH_SHORT).show();
+    }
+
+    @Override
+    public void onEventLongPress(WeekViewEvent event, RectF eventRect) {
+        Toast.makeText(this, "Long pressed event: " + event.getName(), Toast.LENGTH_SHORT).show();
+    }
+
+    @Override
+    public void onEmptyViewLongPress(Calendar time) {
+        AlertDialog.Builder builder = new AlertDialog.Builder(this);
+        View view = LayoutInflater.from(this).inflate(R.layout.assign_menu,null);
+        RelativeLayout today = view.findViewById(R.id.assign_today);
+        RelativeLayout one = view.findViewById(R.id.assign_one);
+        RelativeLayout three = view.findViewById(R.id.assign_three);
+        RelativeLayout week = view.findViewById(R.id.assign_week);
+        today.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                mWeekView.goToToday();
+                setupDateTimeInterpreter(false);
+                dialog.dismiss();
+            }
+        });
+        one.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                if (mWeekViewType != TYPE_DAY_VIEW) {
+                    mWeekViewType = TYPE_DAY_VIEW;
+                    mWeekView.setNumberOfVisibleDays(1);
+
+                    // Lets change some dimensions to best fit the view.
+                    mWeekView.setColumnGap((int) TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_DIP, 8, getResources().getDisplayMetrics()));
+                    mWeekView.setTextSize((int) TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_SP, 12, getResources().getDisplayMetrics()));
+                    mWeekView.setEventTextSize((int) TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_SP, 12, getResources().getDisplayMetrics()));
+                    setupDateTimeInterpreter(false);
+                }
+                dialog.dismiss();
+            }
+        });
+        three.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                if (mWeekViewType != TYPE_THREE_DAY_VIEW) {
+                    mWeekViewType = TYPE_THREE_DAY_VIEW;
+                    mWeekView.setNumberOfVisibleDays(3);
+
+                    // Lets change some dimensions to best fit the view.
+                    mWeekView.setColumnGap((int) TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_DIP, 8, getResources().getDisplayMetrics()));
+                    mWeekView.setTextSize((int) TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_SP, 12, getResources().getDisplayMetrics()));
+                    mWeekView.setEventTextSize((int) TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_SP, 12, getResources().getDisplayMetrics()));
+                    setupDateTimeInterpreter(false);
+                }
+                dialog.dismiss();
+            }
+        });
+        week.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                if (mWeekViewType != TYPE_WEEK_VIEW) {
+                    mWeekViewType = TYPE_WEEK_VIEW;
+                    mWeekView.setNumberOfVisibleDays(7);
+
+                    // Lets change some dimensions to best fit the view.
+                    mWeekView.setColumnGap((int) TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_DIP, 2, getResources().getDisplayMetrics()));
+                    mWeekView.setTextSize((int) TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_SP, 10, getResources().getDisplayMetrics()));
+                    mWeekView.setEventTextSize((int) TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_SP, 10, getResources().getDisplayMetrics()));
+                    setupDateTimeInterpreter(true);
+                }
+                dialog.dismiss();
+            }
+        });
+        builder.setView(view);
+        dialog = builder.show();
+    }
+
+
+    @Override
+    public List<? extends WeekViewEvent> onMonthChange(int newYear, int newMonth) {
+        // Populate the week view with some events.
+
+//        Calendar startTime = Calendar.getInstance();
+//        startTime.set(Calendar.HOUR_OF_DAY, 3);
+//        startTime.set(Calendar.MINUTE, 0);
+//        startTime.set(Calendar.MONTH, newMonth - 1);
+//        startTime.set(Calendar.YEAR, newYear);
+//        Calendar endTime = (Calendar) startTime.clone();
+//        endTime.add(Calendar.HOUR, 1);
+//        endTime.set(Calendar.MONTH, newMonth - 1);
+//        WeekViewEvent event = new WeekViewEvent(1, getEventTitle(startTime), startTime, endTime);
+//        event.setColor(getResources().getColor(R.color.event_color_01));
+//        events.add(event);
+//
+//        startTime = Calendar.getInstance();
+//        startTime.set(Calendar.HOUR_OF_DAY, 3);
+//        startTime.set(Calendar.MINUTE, 30);
+//        startTime.set(Calendar.MONTH, newMonth-1);
+//        startTime.set(Calendar.YEAR, newYear);
+//        endTime = (Calendar) startTime.clone();
+//        endTime.set(Calendar.HOUR_OF_DAY, 4);
+//        endTime.set(Calendar.MINUTE, 30);
+//        endTime.set(Calendar.MONTH, newMonth-1);
+//        event = new WeekViewEvent(10, getEventTitle(startTime), startTime, endTime);
+//        event.setColor(getResources().getColor(R.color.event_color_02));
+//        events.add(event);
+//
+//        startTime = Calendar.getInstance();
+//        startTime.set(Calendar.HOUR_OF_DAY, 4);
+//        startTime.set(Calendar.MINUTE, 20);
+//        startTime.set(Calendar.MONTH, newMonth-1);
+//        startTime.set(Calendar.YEAR, newYear);
+//        endTime = (Calendar) startTime.clone();
+//        endTime.set(Calendar.HOUR_OF_DAY, 5);
+//        endTime.set(Calendar.MINUTE, 0);
+//        event = new WeekViewEvent(10, getEventTitle(startTime), startTime, endTime);
+//        event.setColor(getResources().getColor(R.color.event_color_03));
+//        events.add(event);
+//
+//        startTime = Calendar.getInstance();
+//        startTime.set(Calendar.HOUR_OF_DAY, 5);
+//        startTime.set(Calendar.MINUTE, 30);
+//        startTime.set(Calendar.MONTH, newMonth-1);
+//        startTime.set(Calendar.YEAR, newYear);
+//        endTime = (Calendar) startTime.clone();
+//        endTime.add(Calendar.HOUR_OF_DAY, 2);
+//        endTime.set(Calendar.MONTH, newMonth-1);
+//        event = new WeekViewEvent(2, getEventTitle(startTime), startTime, endTime);
+//        event.setColor(getResources().getColor(R.color.event_color_02));
+//        events.add(event);
+//
+//        startTime = Calendar.getInstance();
+//        startTime.set(Calendar.HOUR_OF_DAY, 5);
+//        startTime.set(Calendar.MINUTE, 0);
+//        startTime.set(Calendar.MONTH, newMonth - 1);
+//        startTime.set(Calendar.YEAR, newYear);
+//        startTime.add(Calendar.DATE, 1);
+//        endTime = (Calendar) startTime.clone();
+//        endTime.add(Calendar.HOUR_OF_DAY, 3);
+//        endTime.set(Calendar.MONTH, newMonth - 1);
+//        event = new WeekViewEvent(3, getEventTitle(startTime), startTime, endTime);
+//        event.setColor(getResources().getColor(R.color.event_color_03));
+//        events.add(event);
+//
+//        startTime = Calendar.getInstance();
+//        startTime.set(Calendar.DAY_OF_MONTH, 15);
+//        startTime.set(Calendar.HOUR_OF_DAY, 3);
+//        startTime.set(Calendar.MINUTE, 0);
+//        startTime.set(Calendar.MONTH, newMonth-1);
+//        startTime.set(Calendar.YEAR, newYear);
+//        endTime = (Calendar) startTime.clone();
+//        endTime.add(Calendar.HOUR_OF_DAY, 3);
+//        event = new WeekViewEvent(4, getEventTitle(startTime), startTime, endTime);
+//        event.setColor(getResources().getColor(R.color.event_color_04));
+//        events.add(event);
+//
+//        startTime = Calendar.getInstance();
+//        startTime.set(Calendar.DAY_OF_MONTH, 1);
+//        startTime.set(Calendar.HOUR_OF_DAY, 3);
+//        startTime.set(Calendar.MINUTE, 0);
+//        startTime.set(Calendar.MONTH, newMonth-1);
+//        startTime.set(Calendar.YEAR, newYear);
+//        endTime = (Calendar) startTime.clone();
+//        endTime.add(Calendar.HOUR_OF_DAY, 3);
+//        event = new WeekViewEvent(5, getEventTitle(startTime), startTime, endTime);
+//        event.setColor(getResources().getColor(R.color.event_color_01));
+//        events.add(event);
+//
+//        startTime = Calendar.getInstance();
+//        startTime.set(Calendar.DAY_OF_MONTH, startTime.getActualMaximum(Calendar.DAY_OF_MONTH));
+//        startTime.set(Calendar.HOUR_OF_DAY, 15);
+//        startTime.set(Calendar.MINUTE, 0);
+//        startTime.set(Calendar.MONTH, newMonth-1);
+//        startTime.set(Calendar.YEAR, newYear);
+//        endTime = (Calendar) startTime.clone();
+//        endTime.add(Calendar.HOUR_OF_DAY, 3);
+//        event = new WeekViewEvent(5, getEventTitle(startTime), startTime, endTime);
+//        event.setColor(getResources().getColor(R.color.event_color_02));
+//        events.add(event);
+//
+//        //AllDay event
+//        startTime = Calendar.getInstance();
+//        startTime.set(Calendar.HOUR_OF_DAY, 0);
+//        startTime.set(Calendar.MINUTE, 0);
+//        startTime.set(Calendar.MONTH, newMonth-1);
+//        startTime.set(Calendar.YEAR, newYear);
+//        endTime = (Calendar) startTime.clone();
+//        endTime.add(Calendar.HOUR_OF_DAY, 23);
+//        event = new WeekViewEvent(7, getEventTitle(startTime),null, startTime, endTime);
+//        event.setColor(getResources().getColor(R.color.event_color_04));
+//        events.add(event);
+//        events.add(event);
+//
+//        startTime = Calendar.getInstance();
+//        startTime.set(Calendar.DAY_OF_MONTH, 8);
+//        startTime.set(Calendar.HOUR_OF_DAY, 2);
+//        startTime.set(Calendar.MINUTE, 0);
+//        startTime.set(Calendar.MONTH, newMonth-1);
+//        startTime.set(Calendar.YEAR, newYear);
+//        endTime = (Calendar) startTime.clone();
+//        endTime.set(Calendar.DAY_OF_MONTH, 10);
+//        endTime.set(Calendar.HOUR_OF_DAY, 23);
+//        event = new WeekViewEvent(8, getEventTitle(startTime),null, startTime, endTime);
+//        event.setColor(getResources().getColor(R.color.event_color_03));
+//        events.add(event);
+//
+//        // All day event until 00:00 next day
+//        startTime = Calendar.getInstance();
+//        startTime.set(Calendar.DAY_OF_MONTH, 10);
+//        startTime.set(Calendar.HOUR_OF_DAY, 0);
+//        startTime.set(Calendar.MINUTE, 0);
+//        startTime.set(Calendar.SECOND, 0);
+//        startTime.set(Calendar.MILLISECOND, 0);
+//        startTime.set(Calendar.MONTH, newMonth-1);
+//        startTime.set(Calendar.YEAR, newYear);
+//        endTime = (Calendar) startTime.clone();
+//        endTime.set(Calendar.DAY_OF_MONTH, 11);
+//        event = new WeekViewEvent(8, getEventTitle(startTime), null, startTime, endTime);
+//        event.setColor(getResources().getColor(R.color.event_color_01));
+//        events.add(event);
+
+        return events;
+    }
+}
